Did some changes in rotary encoder logic and add button function on menu files

+++ Drivers/RotaryEncoder/rotary_encoder.h
@@ -14,7 +14,7 @@ typedef struct {
     uint16_t DT_pin;
     GPIO_TypeDef* SW_port;
     uint16_t SW_pin;
-    volatile int position;
+    volatile float position;
 } RotaryEncoder_TypeDef;
 
 void RotaryEncoder_Init(
